+++ seq2science/scripts/deseq2_rule.R
@@ -12,7 +12,6 @@ threads         <- snakemake@threads[[1]]
 log_file        <- snakemake@log[[1]]
 counts_file     <- snakemake@input[[1]]
 samples_file    <- snakemake@params$samples
-scripts_dir     <- snakemake@params$scripts_dir
 replicates      <- snakemake@params$replicates
 contrast        <- snakemake@wildcards$contrast
 mtp             <- snakemake@config$deseq2$multiple_testing_procedure
@@ -29,26 +28,20 @@ sink(log, type="message")
 
 # log all variables for debugging purposes
 cat('# variables used for this analysis:\n')
-cat('threads      <-',   threads, '\n')
+cat('threads      <- ',  threads,      '\n',  sep = "")
 cat('log_file     <- "', log_file,     '"\n', sep = "")
 cat('counts_file  <- "', counts_file,  '"\n', sep = "")
 cat('samples_file <- "', samples_file, '"\n', sep = "")
-cat('scripts_dir  <- "', scripts_dir,  '"\n', sep = "")
-cat('replicates   <- ',  replicates, '\n')
+cat('replicates   <- ',  replicates,   '\n',  sep = "")
 cat('contrast     <- "', contrast,     '"\n', sep = "")
 cat('mtp          <- "', mtp,          '"\n', sep = "")
-cat('fdr          <-',   fdr, '\n')
+cat('fdr          <- ',  fdr,          '\n',  sep = "")
 cat('se           <- "', se,           '"\n', sep = "")
 cat('assembly     <- "', assembly,     '"\n', sep = "")
-cat('salmon       <- ', salmon,       '\n', sep = "")
+cat('salmon       <- ',  salmon,       '\n',  sep = "")
 cat('output       <- "', output,       '"\n', sep = "")
 cat('\n')
 
 cat('Sessioninfo:\n')
 sessionInfo()
 cat('\n')
-
-
-# run the core script
-deseq_script <- file.path(scripts_dir, "deseq2.R")
-source(deseq_script)